requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.0.1
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2022.1